Introduce single assigment with 'let'
Single-assignment variables are declared using let <iden> = <val>;

Global variables must be single-assignment. Undefined variable access
is usually reported at comptime, but because this is a dynamic language,
globals can be referenced (but not assigned) in local scope
before they are defined. This is only caught at runtime.

Assigning again to a single-assignment variable is always a comptime
error, even for globals.