Strong_typing/Force_emitting_ops_with_correct_operand
Some OpCodes must be followed by an operand,
and some are followed by nothing.

But we want OpCode itself to be 1 byte long,
so we can't use a nested enum. Instead, define
separate enums and From conversions. Then only
allow the compiler to emit enums with their correct
operand count.