Start making Tokens slice into source
Tokens previously copied their lexemes, which becomes detrimental in big
source code. Really, tokens should hols slices into the source code,
and have the same lifetime as the source.

Maybe I'll introduce a copy method later that frees the tokens, and
gives a type that owns its lexeme.

This is not quite working yet, but the idea is right I think